(edit) add the SpeciesView folder and create the ViewBlock sub-component (previously was included in the SpeciesView component file)
